chore: build with vite library mode instead of rollup
Vite 8 bundles with rolldown, so this drops rollup from the tree entirely.

- vite.config.ts -> vite.config.mts, adding build.lib alongside the existing
  vitest config; .mts avoids the ESM-in-CJS config loader warning
- declarations now come from tsc --emitDeclarationOnly, which keeps type
  errors failing the build since rolldown only strips types
- drop rollup, @rollup/plugin-replace, @rollup/plugin-typescript and tslib,
  the last only required as a peer of the typescript plugin
- raise target to es2022, fully supported by the Node 22.12 floor

Output is equivalent to the previous build: identical file names, identical
declarations, the same four external imports, the same 12 exports, require()
rather than import() in the cjs bundle, and PKG_VERSION still inlined.
