+++ package.json
@@ -23,8 +23,8 @@
     "prepublish": "yarn build",
     "test": "vitest",
     "test:bun": "bun run vitest",
-    "start": "rollup -w -c",
-    "build": "rm -rf dist && rollup -c",
+    "start": "vite build --watch",
+    "build": "rm -rf dist && vite build && tsc --emitDeclarationOnly -p tsconfig.json",
     "generate:open-api": "./generate-openapi.sh",
     "lint": "eslint src __tests__ index.ts",
     "lint:fix": "eslint --fix src __tests__ index.ts",
@@ -47,8 +47,6 @@
   ],
   "devDependencies": {
     "@eslint/js": "^10.0.1",
-    "@rollup/plugin-replace": "^6.0.3",
-    "@rollup/plugin-typescript": "^12.3.0",
     "@stream-io/openai-realtime-api": "~0.3.3",
     "@types/node": "^22.20.1",
     "dotenv": "^17.3.1",
@@ -58,8 +56,6 @@
     "husky": "^9.1.7",
     "lint-staged": "^16.3.3",
     "prettier": "^3.8.1",
-    "rollup": "^4.59.0",
-    "tslib": "^2.8.1",
     "typescript": "^5.9.3",
     "typescript-eslint": "^8.67.0",
     "vite": "^8.2.1",
