+++ tsconfig.json
@@ -2,7 +2,7 @@
   "compilerOptions": {
     "outDir": "./dist",
     "module": "ES2020",
-    "target": "ES2020",
+    "target": "ES2022",
     "lib": ["esnext"],
     "noEmitOnError": true,
     "noImplicitAny": true,
