--- vite.config.ts
@@ -1,14 +1,0 @@
-import { defineConfig } from 'vitest/config';
-
-export default defineConfig({
-  build: {},
-  test: {
-    coverage: {
-      reporter: ['lcov'],
-    },
-    testTimeout: 60000,
-    include: ['__tests__/**/*.test.ts'],
-    includeSource: ['src/**/*.ts'],
-    retry: 3,
-  },
-});
